+++ CHANGES.rst
@@ -5,6 +5,18 @@ general
 -------
 - Updated datamodel maker utility imports. [#654]
 
+- Update non-VOunits to using ``astropy.units``. [#658]
+
+- update minimum version of ``asdf`` to ``2.14.2`` and ``jsonschema`` to ``4.0.1`` and added minimum dependency checks to CI [#664]
+
+- Remove use of ``pytest-openfiles`` [#666]
+
+source_detection
+----------------
+- Added SourceDetection Step to pipeline [#608]
+- Added option of fixed random seed for unit tests to avoid intermittent failures from randomness. [#668]
+
+
 0.10.0 (2023-02-21)
 ===================